Added testing for benchmarking run
