+++ testing/scripts/Makefile
@@ -181,6 +181,13 @@ test_notebooks: build_protos install
 		-m "notebooks" \
 		-W ignore test_notebooks.py::TestNotebooks 2>&1
 
+test_benchmark: build_protos install
+	pytest \
+		--verbose \
+		-s \
+		-m "benchmark" \
+		-W ignore test_notebooks.py::TestNotebooks 2>&1
+
 .PHONY: clean
 clean:
 	rm -f proto/prediction*
